+++ src/components/Button2/index.ts
@@ -1,2 +1,0 @@
-export { Button2 } from './Button2';
-export type { Button2Props } from './Button2';